+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## x.x.x (xxxx-xx-xx)
+- **Removed `hello` command**: the terminal window isn't very high so every space counts, so I removed this test command, even though it's a historical command.
+
 ## 0.0.4 (2025-05-18)
 - **Processes**: added the ability to create multiple processes inside the OS
 - **Added PIT support**: PIT is a Programmable Interval Timer, like a clock simulation